gateway request validations added
